+++ tools/integration_tests/monitoring/prom_w_grpc_metrics_test.go
@@ -1,4 +1,4 @@
-// Copyright 2025 Google LLC
+// Copyright 2026 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,11 +18,8 @@ import (
 	"log"
 	"os"
 	"path"
-	"strings"
 	"testing"
 
-	"github.com/googlecloudplatform/gcsfuse/v3/tools/integration_tests/util/client"
-	"github.com/googlecloudplatform/gcsfuse/v3/tools/integration_tests/util/operations"
 	"github.com/googlecloudplatform/gcsfuse/v3/tools/integration_tests/util/setup"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -31,30 +28,6 @@ import (
 // PromGrpcMetricsTest is the test suite for gRPC metrics.
 type PromGrpcMetricsTest struct {
 	PromTestBase
-	flags          []string
-	prometheusPort int
-}
-
-func (p *PromGrpcMetricsTest) SetupSuite() {
-	setup.SetUpLogFilePath("TestPromGrpcMetricsSuite", gkeTempDir, "", testEnv.cfg)
-	mountGCSFuseAndSetupTestDir(p.flags, testEnv.ctx, testEnv.storageClient)
-}
-
-func (p *PromGrpcMetricsTest) TearDownSuite() {
-	setup.UnmountGCSFuseWithConfig(testEnv.cfg)
-}
-
-func (p *PromGrpcMetricsTest) SetupTest() {
-	// Create a new directory for each test.
-	testName := strings.ReplaceAll(p.T().Name(), "/", "_")
-	gcsDir := path.Join(testDirName, testName)
-	testEnv.testDirPath = path.Join(mountDir, gcsDir)
-	operations.CreateDirectory(testEnv.testDirPath, p.T())
-	client.SetupFileInTestDirectory(testEnv.ctx, testEnv.storageClient, gcsDir, "hello.txt", 10, p.T())
-}
-
-func (p *PromGrpcMetricsTest) TearDownTest() {
-	setup.SaveGCSFuseLogFileInCaseOfFailure(p.T())
 }
 
 func (p *PromGrpcMetricsTest) TestStorageClientGrpcMetrics() {
@@ -76,6 +49,7 @@ func (p *PromGrpcMetricsTest) TestStorageClientGrpcMetrics() {
 
 func TestPromGrpcMetricsSuite(t *testing.T) {
 	ts := &PromGrpcMetricsTest{}
+	ts.suiteName = "TestPromGrpcMetricsSuite"
 	if testEnv.cfg.GKEMountedDirectory == "" {
 		// Skip the test if the testing environment is GCE VM.
 		t.SkipNow()
